backend/requirements.txt
asgiref==3.11.1
Django==5.0.7
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2
psycopg2-binary==2.9.11
python-dotenv==1.2.1
sqlparse==0.5.5
tzdata==2025.3
boto3==1.35.96
django-storages==1.14.4
celery==5.3.4
redis==5.0.1
requests==2.31.0
yookassa==3.3.0
whitenoise==6.7.0
Pillow==10.4.0
sentry-sdk[django]==2.19.2
python-magic==0.4.27
python-json-logger==2.0.7
